requirements.txt
aiofiles==0.7.0
anyio==3.6.1
asgiref==3.4.1
astroid==2.8.0
attrs==22.1.0
bcrypt==3.2.2
bson==0.5.10
cffi==1.15.1
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
isort==5.9.3
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
more-itertools==8.13.0
motor==3.0.0
packaging==21.3
passlib==1.7.4
pkg-resources==0.0.0
platformdirs==2.4.0
pluggy==0.13.1
py==1.11.0
pycparser==2.21
pydantic==1.9.1
pylint==2.11.1
pymongo==4.2.0
pyparsing==3.0.9
pytest==5.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==4.3.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.8.1
